Move log previous crash method to base class, the only difference here was the Enum for special folder, both of which resolve to the same value.
