+++ Com.Kumulos.Android/KumulosImplementation.cs
@@ -176,7 +176,7 @@ namespace Com.Kumulos
             Android.Kumulos.PushUnregister(Application.Context.ApplicationContext);
         }
 
-        public void TrackEvent(string eventType, Dictionary<string, object> properties)
+        public override void TrackEvent(string eventType, Dictionary<string, object> properties)
         {
             JSONObject props = new JSONObject(properties);
             Android.Kumulos.TrackEvent(Application.Context.ApplicationContext, eventType, props);
@@ -264,42 +264,7 @@ namespace Com.Kumulos
             File.WriteAllText(filename, JsonConvert.SerializeObject(crash, Formatting.None));
         }
 
-        private void LogPreviousCrash()
-        {
-            var documents = Environment.GetFolderPath(Environment.SpecialFolder.Personal);
-            var filename = Path.Combine(documents, "CrashLog.json");
-
-            if (!File.Exists(filename))
-            {
-                return;
-            }
-
-            var text = File.ReadAllText(filename);
-            var jsonObj = (JContainer)JsonConvert.DeserializeObject(text);
-
-            var dict = new Dictionary<string, object>
-                {
-                    { "format", (string)jsonObj["format"] },
-                    { "uncaught", (bool)jsonObj["uncaught"] }
-                };
-
-            var reportObj = (JContainer)jsonObj["report"];
-
-            var report = new Dictionary<string, object>
-            {
-                { "stackTrace", (string)reportObj["stackTrace"] },
-                { "message", (string)reportObj["message"] },
-                { "type", (string)reportObj["type"] },
-                { "source", (string)reportObj["source"] },
-                { "lineNumber", (int)reportObj["lineNumber"] }
-            };
-
-            dict.Add("report", report);
-
-            TrackEvent(Consts.CRASH_REPORT_EVENT_TYPE, dict);
-
-            File.Delete(filename);
-        }
+       
 
         public void TrackiBeaconProximity(object CLBeaconObject)
         {
